refactor: move auth inside lib; fix: persist script
